move loadPublishedDataset middleware
