+++ src/controllers/consumer.ts
@@ -7,7 +7,6 @@ import { DatasetListItemDTO } from '../dtos/dataset-list-item-dto';
 import { UnknownException } from '../exceptions/unknown.exception';
 import { DatasetRepository } from '../repositories/dataset';
 import { NotFoundException } from '../exceptions/not-found.exception';
-import { hasError, datasetIdValidator } from '../validators';
 import { ConsumerDatasetDTO } from '../dtos/consumer-dataset-dto';
 import { DownloadFormat } from '../enums/download-format';
 import { BadRequestException } from '../exceptions/bad-request.exception';
@@ -37,29 +36,6 @@ export const listPublishedDatasets = async (req: Request, res: Response, next: N
         logger.error(err, 'Failed to fetch published dataset list');
         next(new UnknownException());
     }
-};
-
-export const loadPublishedDataset = async (req: Request, res: Response, next: NextFunction) => {
-    const datasetIdError = await hasError(datasetIdValidator(), req);
-
-    if (datasetIdError) {
-        logger.error(datasetIdError);
-        next(new NotFoundException('errors.dataset_id_invalid'));
-        return;
-    }
-
-    try {
-        logger.debug(`Loading published dataset ${req.params.dataset_id}...`);
-        const dataset = await DatasetRepository.getPublishedById(req.params.dataset_id);
-        res.locals.datasetId = dataset.id;
-        res.locals.dataset = dataset;
-    } catch (err) {
-        logger.error(`Failed to load dataset, error: ${err}`);
-        next(new NotFoundException('errors.no_dataset'));
-        return;
-    }
-
-    next();
 };
 
 export const getPublishedDatasetById = async (req: Request, res: Response, next: NextFunction) => {
